chore: tidy scripts
