+++ apps/single-app/app/package.json
@@ -6,7 +6,7 @@
     "scripts": {
         "build": "rimraf dist && yarn build:main",
         "build:main": "tsc -p tsconfig.json",
-        "build-win32": "mkdir deploy & node ../../../scripts/build-win32.js package-manager-single-app.exe && node ../../../scripts/copy-natives.js win32-x64",
+        "build-win32": "mkdir deploy & node ../../../scripts/build-win32.mjs package-manager-single-app.exe && node ../../../scripts/copy-natives.js win32-x64",
         "__test": "jest",
         "start": "node --inspect dist/index.js"
     },
