+++ apps/worker/app/package.json
@@ -7,7 +7,7 @@
         "build": "rimraf dist && yarn build:main",
         "build:main": "tsc -p tsconfig.json",
         "oldbuild-win32": "mkdir deploy & rimraf deploy/worker.exe  && nexe dist/index.js -t windows-x64-12.18.1 -o deploy/worker.exe && node scripts/copy-natives.js win32-x64",
-        "build-win32": "mkdir deploy & node ../../../scripts/build-win32.js worker.exe && node ../../../scripts/copy-natives.js win32-x64",
+        "build-win32": "mkdir deploy & node ../../../scripts/build-win32.mjs worker.exe && node ../../../scripts/copy-natives.js win32-x64",
         "__test": "jest",
         "start": "node dist/index.js"
     },
