+++ package.json
@@ -28,8 +28,8 @@
         "test:update:changed": "lerna run --since origin/master --include-dependents test -- -u",
         "typecheck": "lerna exec -- tsc --noEmit",
         "typecheck:changed": "lerna exec --since origin/master --include-dependents -- tsc --noEmit",
-        "build-win32": "node scripts/prepare-for-build32.js && lerna run build-win32 --stream && node scripts/cleanup-after-build32.js",
-        "gather-built": "node scripts/gather-all-built.js",
+        "build-win32": "node scripts/prepare-for-build32.js && lerna run build-win32 --stream && node scripts/cleanup-after-build32.mjs",
+        "gather-built": "node scripts/gather-all-built.mjs",
         "sign-executables": "node scripts/sign-executables.mjs",
         "start:http-server": "lerna run start --stream --scope @http-server/app",
         "start:workforce": "lerna run start --stream --scope @workforce/app",
