+++ scripts/build-win32.mjs
@@ -1,17 +1,16 @@
-/* eslint-disable node/no-unpublished-require, node/no-extraneous-require */
+/* eslint-disable node/no-unpublished-import, node/no-extraneous-import, no-console */
 
-const promisify = require('util').promisify
-const cp = require('child_process')
-const path = require('path')
-const nexe = require('nexe')
+import { promisify } from 'util'
+import cp from 'child_process'
+import path from 'path'
+import nexe from 'nexe'
+import glob0 from 'glob'
+import fse from 'fs-extra'
+import { createRequire } from 'module'
+
 const exec = promisify(cp.exec)
-const glob = promisify(require('glob'))
-
-const fse = require('fs-extra')
-const mkdirp = require('mkdirp')
-const rimraf = promisify(require('rimraf'))
-
-const fseCopy = promisify(fse.copy)
+const glob = promisify(glob0)
+const require = createRequire(import.meta.url)
 
 /*
 	Due to nexe not taking into account the packages in the mono-repo, we're doing a hack,
@@ -26,7 +25,6 @@ if (!executableName) {
 }
 
 ;(async () => {
-
 	log(`Collecting dependencies for ${packageJson.name}...`)
 	// List all Lerna packages:
 	const list = await exec('yarn lerna list -a --json')
@@ -34,7 +32,7 @@ if (!executableName) {
 
 	const packages = JSON.parse(str)
 
-	await mkdirp(basePath + 'node_modules')
+	await fse.mkdirp(basePath + 'node_modules')
 
 	// Copy the packages into node_modules:
 	const copiedFolders = []
@@ -49,7 +47,7 @@ if (!executableName) {
 		const target = path.resolve(path.join(basePath, 'node_modules', package0.name))
 
 		// log(`    ${source} -> ${target}`)
-		ps.push(fseCopy(source, target))
+		ps.push(fse.copy(source, target))
 
 		copiedFolders.push(target)
 	}
@@ -71,7 +69,7 @@ if (!executableName) {
 			!file.match(/node_modules$/) &&
 			!file.match(/dist$/)
 		) {
-			ps.push(rimraf(file))
+			ps.push(fse.rm(file, { recursive: true }))
 		}
 	}
 	await Promise.all(ps)
@@ -93,7 +91,7 @@ if (!executableName) {
 	log(`Cleaning up...`)
 	// Clean up after ourselves:
 	for (const copiedFolder of copiedFolders) {
-		await rimraf(copiedFolder)
+		await fse.rm(copiedFolder, { recursive: true })
 	}
 
 	log(`...done!`)
